Upload Bare Functionality
Add bare functions for opening a file selection dialog, and filtering to include only
images. Once an image is selected, it is sent to the plantnet api to identify the
plant species and identification score. Additional information is available but not
currently extracted from the JSONObject received. All logic is currently accumulated
in the UploadInteractor class, and awaits distribution to appropriate methods and
classes to adhere to Clean Architecture. Several exceptions, edge cases, and user
behaviours are not accounted for. It remains to call the trefle.io api for additional
information on the uploaded plant.