Upload Main View Setup
Add rudimentary demo for the main view, from which the upload use case may be
triggered by a button press. Once the button is pressed, an in-window pop-up view
has been set up. All logic remains in the UploadInteractor class, but it is now
possible to start distributing logic among files that adhere to Clean Architecture.