Adopt Clean Architecture (Step 1)
Clean Architecture classes are introduced, but not completed. There is one
interactor, presenter, and controller, and two view models and views. An additional
view may be added in the near future. If the current implementation of only one of
each interface adapter and interactor is evaluated to violate the single
responsibility principle, it will be separated into more classes. The program opens
an upload overlay when the upload button is clicked, then allows the user to select
an image file from their device. Currently the file path is extracted and sent to
the Presenter, where nothing is yet done with it. There is currently a design issue
where it is unclear how to switch the panel in the cardlayout once a new upload
view should be displayed.