Add Confirm View after Upload
The UploadConfirmView class is implemented. The user is now greeted with a display
of their image upon selecting from the file explorer. This image is cropped to a
square 500x500 view. The user then has the option to return to the file selection
view to choose a new file, or confirm and go to the results view. Both corresponding
buttons work. The UploadResultView class and its corresponding methods throughout
the Upload interactor and interface adapters remain.

There are currently two bugs identified in this system. They are reported as a TODO
comment within the MainViewDemo class.