Fix: Button Style Alterations
The previously reported bug where button styling is altered after attempting an upload
has been patched. The issue regarded restyling the Look and Feel of the application to
produce a file selection dialog that resembled the user's system. The Look and Feel
is now returned to the default after closing the file selection dialog.