Select & Confirm Styling
Panel design for UploadSelectView and UploadConfirmView is added, so that the first
two views are stylized. Additionally, the Main View is now also expected to
implement disableInteraction and enableInteraction methods, to ensure that the user
does not click any buttons outside the overlay while a popup is active. These
methods are not used within the use case logic, but instead are called upon in the
Main View overlay method as helpers.