Result View Stylization
The result view is now designed in a pleasant way so that it displays the uploaded
image, API information drawn from that image, and user notes. The user notes are
customizable. The view is currently a mock-up, in that it is not connected to the
interactor and does not receive photo-specific information from the API. Next steps
include connecting the interactor, and introducing save button functionality.