Port Upload to MainView
The upload use case has been ported to the MainView class. This involves changing
the MainView to a JLayeredPane to allow for overlays. All entities that were already
added to the main view are now added to a buffer mainPanel, which is then added to
the default layer of the MainView class. The overlay, enableInteraction, and
disableInteraction methods are transferred from the MainViewDemo. overlayUploadView()
is also added, but is likely to be a temporary setup for the upload use case.