Plant Edit Foundations
The overlay within which the user may view an existing image that they own is now
developed. This needs to be integrated into the main method. Currently, it relies
on the editPlant method in the data access object, with is not implemented. Further,
additional fields within the database are necessary in order to retrieve all relevant
information about the plant.